Adding perception/semantic_parsing tests
new preprocess tests

fix model and data path

refactor locobot test + clean up base_util

fix model and data path
