+++ agents/craftassist/tests/test_with_model.py
@@ -12,8 +12,8 @@ from droidlet.lowlevel.minecraft.mc_util import euclid_dist
 from agents.craftassist.tests.base_craftassist_test_case import BaseCraftassistTestCase
 from droidlet.shared_data_structs import MockOpt
 
-TTAD_MODEL_DIR = os.path.join(os.path.dirname(__file__), "../../../../agents/craftassist/models/semantic_parser/")
-TTAD_BERT_DATA_DIR = os.path.join(os.path.dirname(__file__), "../../../../agents/craftassist/datasets/annotated_data/")
+TTAD_MODEL_DIR = os.path.join(os.path.dirname(__file__), "../models/semantic_parser/")
+TTAD_BERT_DATA_DIR = os.path.join(os.path.dirname(__file__), "../datasets/annotated_data/")
 
 
 class PutMemoryTestCase(BaseCraftassistTestCase):
